requirements.txt
lazy-object-proxy==1.3.1
graphene-django==2.0.0
pylint-django==0.7.2
snowballstemmer==1.2.1
graphene==2.0
backports.functools-lru-cache==1.2.1
promise==2.1
configparser==3.5.0
requirements-detector==0.5.2
wrapt==1.10.11
mccabe==0.6.1
pylint-celery==0.3
psycopg2==2.7.3.2
six==1.11.0
rx==1.6.0
pylint-common==0.2.5
djangorestframework==3.7.1
isort==4.2.15
typing==3.6.2
olefile==0.44
pep8-naming==0.4.1
arrow==0.12.0
pylint==1.7.4
urllib3==1.22
iso8601==0.1.12
pytz==2017.3
python-dateutil==2.6.1
django-graphiql==0.4.4
dodgy==0.1.9
prospector==0.12.7
astroid==1.5.3
pyyaml==3.12
certifi==2017.11.5
graphql-core==2.0
graphql-relay==0.4.5
enum34==1.1.6
pydocstyle==2.1.1
pyflakes==1.6.0
setoptconf==0.2.0
django==1.11.7
pylint-flask==0.5
chardet==3.0.4
pylint-plugin-utils==0.2.6
pycodestyle==2.0.0
singledispatch==3.4.0.3
requests==2.18.4
idna==2.6
pillow==4.3.0
cloudinary==1.11.0
beautifulsoup4==4.6.0
googlemaps==2.5.1
django-user-accounts==2.0.3
social-auth-app-django==2.1.0
django-paypal==0.5.0
django-filer==1.3.2
git+https://github.com/apixu/apixu-python.git
bitpay==2.3.3